add refl parameter redefine device, move refl  plans out to general refl area
